+++ CHANGELOG.md
@@ -240,7 +240,7 @@ Here's the most notable changes since `v0.10.0`. Its been about a year since a r
 - **Migrated to `electron-builder@4`**
 - **Migrated to `webpack@2`**
 - **Migrated to `react-hot-loader@3`**
-- **Changed default live reload server PORT to `3006` from `3000`**
+- **Changed default live reload server PORT to `3007` from `3000`**
 
 #### Additions